+++ src/components/steps/step.vue
@@ -3,8 +3,8 @@
         <div :class="[prefixCls + '-tail']"><i></i></div>
         <div :class="[prefixCls + '-head']">
             <div :class="[prefixCls + '-head-inner']">
-                <span v-if="!icon && currentStatus != 'finish' && currentStatus != 'error'">{{ stepNumber }}</span>
-                <span v-else :class="iconClasses"></span>
+                <span v-if="!icon && currentStatus != 'finish' && currentStatus != 'error'" :title="tips">{{ stepNumber }}</span>
+                <span v-else :class="iconClasses" :title="tips"></span>
             </div>
         </div>
         <div :class="[prefixCls + '-main']">
@@ -30,6 +30,10 @@
                 validator (value) {
                     return oneOf(value, ['wait', 'process', 'finish', 'error']);
                 }
+            },
+            tips: {
+                type: String,
+                default: ''
             },
             title: {
                 type: String,